correct method of maintaining BC with psr-0 classes.
