--- src/lib/legacy/Zikula/Module/CategoriesModule/Entity/CategoryEntity.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace Zikula\Module\CategoriesModule\Entity;
-
-/**
- * @deprecated remove at Core-2.0
- * @see Zikula\CategoriesModule\Entity\CategoryEntity
- *
- * This class is necessary because of the refactoring to psr-4
- * This class maintains the 1.4.x BC API
-
- * Class CategoryEntity
- * @package Zikula\Module\CategoriesModule\Entity
- */
-class CategoryEntity extends \Zikula\CategoriesModule\Entity\CategoryEntity
-{
-}
