--- src/lib/legacy/Zikula/Module/UsersModule/Entity/UserEntity.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace Zikula\Module\UsersModule\Entity;
-
-/**
- * @deprecated remove at Core-2.0
- * @see Zikula\UsersModule\Entity\UserEntity
- *
- * This class is necessary because of the refactoring to psr-4
- * This class maintains the 1.4.x BC API
-
- * Class UserEntity
- * @package Zikula\Module\UsersModule\Entity
- */
-class UserEntity extends \Zikula\UsersModule\Entity\UserEntity
-{
-}
